+++ gazebo/rendering/Camera.hh
@@ -14,9 +14,8 @@
  * limitations under the License.
  *
 */
-
-#ifndef _GAZEBO_RENDERING_CAMERA_HH_
-#define _GAZEBO_RENDERING_CAMERA_HH_
+#ifndef GAZEBO_RENDERING_CAMERA_HH_
+#define GAZEBO_RENDERING_CAMERA_HH_
 
 #include <memory>
 #include <functional>
@@ -400,6 +399,43 @@ namespace gazebo
 
       /// \brief Capture data once and save to disk
       public: void SetCaptureDataOnce();
+
+      /// \brief Turn on video recording.
+      /// \param[in] _format String that represents the video type.
+      /// Supported types include: "avi", "ogv", mp4", "v4l2". If using
+      /// "v4l2", you must also specify a _filename.
+      /// \param[in] _filename Name of the file that stores the video while it
+      /// is being created. This is a temporary file when recording to
+      /// disk, or a video4linux loopback device like /dev/video0 when
+      /// the _format is "v4l2". If blank, a default temporary file is used.
+      /// However, the "v4l2" _format must be accompanied with a video
+      /// loopback device filename.
+      /// \return True on success. The return value is set by
+      /// common::VideoEncoder::Start().
+      /// \sa common::VideoEncoder::Start
+      public: bool StartVideo(const std::string &_format,
+                              const std::string &_filename = "");
+
+      /// \brief Turn off video recording
+      /// \return True on success. The return value is set by
+      /// common::VideoEncoder::Stop().
+      /// \sa common::VideoEncoder::Stop
+      public: bool StopVideo();
+
+      /// \brief Save the last encoded video to disk
+      /// \param[in] _filename File in which to save the encoded video
+      /// \return True if saving was successful. The return value is set by
+      /// common::VideoEncoder::SaveToFile().
+      /// \sa common::VideoEncoder::SaveToFile
+      public: bool SaveVideo(const std::string &_filename);
+
+      /// \brief Reset video recording. This will call
+      /// common::VideoEncoder::Reset, which will cleanup temporary files and
+      /// set video encoding values to their default settings.
+      /// \sa common::VideoEncoder::Reset
+      /// \return True if reset was succesful. Currently this function will
+      /// always return true.
+      public: bool ResetVideo();
 
       /// \brief Set the render target
       /// \param[in] _textureName Name of the new render texture