+++ Changelog.md
@@ -2,6 +2,12 @@
 
 ## Gazebo 8.x.x (2017-xx-xx)
 
+1. Add video recording of user camera. This change added an optional
+   dependency on libavdevice>=56.4.100 for linux systems. When installed,
+   libavdevice will allow a user to stream a simulated camera to a video4linux2 
+   loopback device.
+    * [Pull request #2443](https://bitbucket.org/osrf/gazebo/pull-request/2443)
+
 1. Removed deprecations
     * [Pull request #2427]((https://bitbucket.org/osrf/gazebo/pull-request/2427)
 
@@ -12,6 +18,7 @@
     * [Pull request #2348](https://bitbucket.org/osrf/gazebo/pull-request/2348)
     * [Pull request #2325](https://bitbucket.org/osrf/gazebo/pull-request/2325)
     * [Pull request #2382](https://bitbucket.org/osrf/gazebo/pull-request/2382)
+    * [Pull request #2448](https://bitbucket.org/osrf/gazebo/pull-request/2448)
 
 1. Renamed `gazebo/gui/SaveDialog` to `gazebo/gui/SaveEntityDialog`. A new
    `SaveDialog` class will be added in a future pull request. The migration
@@ -159,10 +166,12 @@
 
 ## Gazebo 7.x.x (2016-xx-xx)
 
+1. Add support for loading wavefront obj mesh files
+    * [Pull request 2454](https://bitbucket.org/osrf/gazebo/pull-request/2454)
+
 1. Added filesystem operations to the common library. Additions include
    `cwd`, `exists`, `isDirectory`, `isFile`, `copyFile`, and `moveFile`.
     * [Pull request 2417](https://bitbucket.org/osrf/gazebo/pull-request/2417)
-
 
 1. Fix loading collada files with multiple texture coordinates.
     * [Pull request 2413](https://bitbucket.org/osrf/gazebo/pull-request/2413)